server/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
passlib[bcrypt]
python-jose
email-validator
motor
